requirements.txt
typer
setuptools
tabulate
python-a2s
vdf
typer
setuptools